Testing changes Tests moved from entries to value_type Additional test cases in io & setup_vars
