+++ src/settings/entries.rs
@@ -187,181 +187,88 @@ impl FTLConfEntry {
 
 #[cfg(test)]
 mod tests {
-    use super::{FTLConfEntry, SetupVarsEntry};
+    use super::{FTLConfEntry, SetupVarsEntry, ValueType};
 
     #[test]
-    fn test_validate_setup_vars_valid() {
+    fn test_ftlconf_value_types() {
         let tests = vec![
-            // Valid parameters
-            (SetupVarsEntry::ApiQueryLogShow, "all", true),
-            (SetupVarsEntry::ApiPrivacyMode, "false", true),
-            (SetupVarsEntry::DnsBogusPriv, "true", true),
-            (SetupVarsEntry::DnsFqdnRequired, "true", true),
-            (SetupVarsEntry::ConditionalForwarding, "true", true),
-            (SetupVarsEntry::ConditionalForwardingDomain, "hub", true),
-            (SetupVarsEntry::ConditionalForwardingIp, "192.168.1.1", true),
+            (FTLConfEntry::AaaaQueryAnalysis, ValueType::YesNo),
+            (FTLConfEntry::BlockingMode, {
+                ValueType::String(&["NULL", "IP-AAAA-NODATA", "IP", "NXDOMAIN"])
+            }),
+            (FTLConfEntry::DbFile, ValueType::Pathname),
+            (FTLConfEntry::DbInterval, ValueType::Decimal),
+            (FTLConfEntry::FtlPort, ValueType::PortNumber),
+            (FTLConfEntry::IgnoreLocalHost, ValueType::YesNo),
+            (FTLConfEntry::MaxDbDays, ValueType::Integer),
+            (FTLConfEntry::MaxLogAge, ValueType::Decimal),
+            (
+                FTLConfEntry::PrivacyLevel,
+                ValueType::String(&["0", "1", "2", "3"])
+            ),
+            (FTLConfEntry::QueryDisplay, ValueType::YesNo),
+            (FTLConfEntry::ResolveIpv4, ValueType::YesNo),
+            (FTLConfEntry::ResolveIpv6, ValueType::YesNo),
+            (
+                FTLConfEntry::SocketListening,
+                ValueType::String(&["localonly", "all"])
+            ),
+        ];
+
+        for (entry, valuetype) in tests {
+            assert_eq!(entry.value_type(), valuetype);
+        }
+    }
+
+    #[test]
+    fn test_setupvars_value_types() {
+        let tests = vec![
+            (
+                SetupVarsEntry::ApiQueryLogShow,
+                ValueType::String(&["all", ""])
+            ),
+            (SetupVarsEntry::ApiPrivacyMode, ValueType::Boolean),
+            (SetupVarsEntry::DnsBogusPriv, ValueType::Boolean),
+            (SetupVarsEntry::DnsFqdnRequired, ValueType::Boolean),
+            (SetupVarsEntry::ConditionalForwarding, ValueType::Boolean),
+            (
+                SetupVarsEntry::ConditionalForwardingDomain,
+                ValueType::Domain
+            ),
+            (SetupVarsEntry::ConditionalForwardingIp, ValueType::Ipv4),
             (
                 SetupVarsEntry::ConditionalForwardingReverse,
-                "1.168.192.in-addr.arpa",
-                true
+                ValueType::ConditionalForwardingReverse
             ),
-            (SetupVarsEntry::DhcpActive, "false", true),
-            (SetupVarsEntry::DhcpEnd, "199.199.1.255", true),
-            (SetupVarsEntry::DhcpIpv6, "false", true),
-            (SetupVarsEntry::DhcpLeasetime, "24", true),
-            (SetupVarsEntry::DhcpStart, "199.199.1.0", true),
-            (SetupVarsEntry::DhcpRouter, "192.168.1.1", true),
-            (SetupVarsEntry::DnsmasqListening, "all", true),
-            (SetupVarsEntry::Dnssec, "false", true),
-            (SetupVarsEntry::InstallWebServer, "true", true),
-            (SetupVarsEntry::InstallWebInterface, "true", true),
-            (SetupVarsEntry::Ipv4Address, "192.168.1.205/24", true),
+            (SetupVarsEntry::DhcpActive, ValueType::Boolean),
+            (SetupVarsEntry::DhcpEnd, ValueType::Ipv4),
+            (SetupVarsEntry::DhcpIpv6, ValueType::Boolean),
+            (SetupVarsEntry::DhcpLeasetime, ValueType::Integer),
+            (SetupVarsEntry::DhcpStart, ValueType::Ipv4),
+            (SetupVarsEntry::DhcpRouter, ValueType::Ipv4),
             (
-                SetupVarsEntry::Ipv6Address,
-                "2001:470:66:d5f:114b:a1b9:2a13:c7d9",
-                true
+                SetupVarsEntry::DnsmasqListening,
+                ValueType::String(&["all", "lan", "single", ""])
             ),
-            (SetupVarsEntry::PiholeDns(0), "8.8.4.4", true),
-            (SetupVarsEntry::PiholeDomain, "lan", true),
-            (SetupVarsEntry::PiholeInterface, "lo", true),
-            (SetupVarsEntry::QueryLogging, "true", true),
-            (SetupVarsEntry::WebUiBoxedLayout, "boxed", true),
-            (SetupVarsEntry::WebEnabled, "false", true),
-        ];
-
-        for (setting, value, result) in tests {
-            assert_eq!(
-                setting.is_valid(value),
-                result,
-                "{:?}.is_valid({:?}) == {}",
-                setting,
-                value,
-                result
-            );
-        }
-    }
-
-    #[test]
-    fn test_validate_setup_vars_invalid() {
-        let tests = vec![
-            // Valid parameters
-            (SetupVarsEntry::ApiQueryLogShow, "41", false),
-            (SetupVarsEntry::ApiPrivacyMode, "off", false),
-            (SetupVarsEntry::DnsBogusPriv, "on", false),
-            (SetupVarsEntry::DnsFqdnRequired, "1", false),
-            (SetupVarsEntry::ConditionalForwarding, "disabled", false),
-            (SetupVarsEntry::ConditionalForwardingDomain, "%%@)#", false),
-            (SetupVarsEntry::ConditionalForwardingIp, "192.1.1", false),
+            (SetupVarsEntry::Dnssec, ValueType::Boolean),
+            (SetupVarsEntry::InstallWebServer, ValueType::Boolean),
+            (SetupVarsEntry::InstallWebInterface, ValueType::Boolean),
+            (SetupVarsEntry::Ipv4Address, ValueType::Ipv4Mask),
+            (SetupVarsEntry::Ipv6Address, ValueType::Ipv6),
+            (SetupVarsEntry::PiholeDns(7), ValueType::Ipv4),
+            (SetupVarsEntry::PiholeDomain, ValueType::Domain),
+            (SetupVarsEntry::PiholeInterface, ValueType::Interface),
+            (SetupVarsEntry::QueryLogging, ValueType::Boolean),
+            (SetupVarsEntry::WebEnabled, ValueType::Boolean),
+            (SetupVarsEntry::WebPassword, ValueType::WebPassword),
             (
-                SetupVarsEntry::ConditionalForwardingReverse,
-                "in-addr.arpa.1.1.1",
-                false
+                SetupVarsEntry::WebUiBoxedLayout,
+                ValueType::String(&["boxed", ""])
             ),
-            (SetupVarsEntry::DhcpActive, "active", false),
-            (
-                SetupVarsEntry::DhcpEnd,
-                "2001:470:66:d5f:114b:a1b9:2a13:c7d9",
-                false
-            ),
-            (SetupVarsEntry::DhcpIpv6, "ipv4", false),
-            (SetupVarsEntry::DhcpLeasetime, "hours", false),
-            (SetupVarsEntry::DhcpStart, "199199.1.0", false),
-            (SetupVarsEntry::DhcpRouter, "192.1681.1", false),
-            (SetupVarsEntry::DnsmasqListening, "dnsmasq", false),
-            (SetupVarsEntry::Dnssec, "1", false),
-            (SetupVarsEntry::InstallWebServer, "yes", false),
-            (SetupVarsEntry::InstallWebInterface, "no", false),
-            (SetupVarsEntry::Ipv4Address, "192.168.1.205", false),
-            (SetupVarsEntry::Ipv6Address, "192.168.1.205", false),
-            (SetupVarsEntry::PiholeDns(0), "www.pi-hole.net", false),
-            (SetupVarsEntry::PiholeDomain, "too many words", false),
-            (SetupVarsEntry::PiholeInterface, "/dev/net/eth1", false),
-            (SetupVarsEntry::QueryLogging, "disabled", false),
-            (SetupVarsEntry::WebUiBoxedLayout, "true", false),
-            (SetupVarsEntry::WebEnabled, "457", false),
         ];
 
-        for (setting, value, result) in tests {
-            assert_eq!(
-                setting.is_valid(value),
-                result,
-                "{:?}.is_valid({:?}) == {}",
-                setting,
-                value,
-                result
-            );
-        }
-    }
-
-    #[test]
-    fn test_validate_setup_vars_disabled() {
-        // Setting the web password is not allowed - must report false.
-        assert_eq!(
-            SetupVarsEntry::WebPassword
-                .is_valid("B350486529B6022919491965A235157110B12437514315201184143ABBB37A14"),
-            false
-        );
-    }
-
-    #[test]
-    fn test_validate_ftl_config_valid() {
-        let tests = vec![
-            // Valid values
-            (FTLConfEntry::AaaaQueryAnalysis, "no", true),
-            (FTLConfEntry::BlockingMode, "NULL", true),
-            (FTLConfEntry::DbInterval, "5.0", true),
-            (FTLConfEntry::DbFile, "/etc/test.conf", true),
-            (FTLConfEntry::FtlPort, "64738", true),
-            (FTLConfEntry::IgnoreLocalHost, "yes", true),
-            (FTLConfEntry::MaxDbDays, "3", true),
-            (FTLConfEntry::MaxLogAge, "8", true),
-            (FTLConfEntry::PrivacyLevel, "2", true),
-            (FTLConfEntry::QueryDisplay, "yes", true),
-            (FTLConfEntry::ResolveIpv6, "yes", true),
-            (FTLConfEntry::ResolveIpv4, "no", true),
-            (FTLConfEntry::SocketListening, "localonly", true),
-        ];
-
-        for (setting, value, result) in tests {
-            assert_eq!(
-                setting.is_valid(value),
-                result,
-                "{:?}.is_valid({:?}) == {}",
-                setting,
-                value,
-                result
-            );
-        }
-    }
-
-    #[test]
-    fn test_validate_ftl_conf_invalid() {
-        let tests = vec![
-            // Invalid values
-            (FTLConfEntry::AaaaQueryAnalysis, "", false),
-            (FTLConfEntry::BlockingMode, "enabled", false),
-            (FTLConfEntry::DbInterval, "true", false),
-            (FTLConfEntry::DbFile, "FTL.conf", false),
-            (FTLConfEntry::FtlPort, "65537", false),
-            (FTLConfEntry::IgnoreLocalHost, "OK", false),
-            (FTLConfEntry::MaxDbDays, "null", false),
-            (FTLConfEntry::MaxLogAge, "enabled", false),
-            (FTLConfEntry::PrivacyLevel, ">9000", false),
-            (FTLConfEntry::QueryDisplay, "disabled", false),
-            (FTLConfEntry::ResolveIpv6, "true", false),
-            (FTLConfEntry::ResolveIpv4, "false", false),
-            (FTLConfEntry::SocketListening, "eth0", false),
-        ];
-
-        for (setting, value, result) in tests {
-            assert_eq!(
-                setting.is_valid(value),
-                result,
-                "{:?}.is_valid({:?}) == {}",
-                setting,
-                value,
-                result
-            );
+        for (entry, valuetype) in tests {
+            assert_eq!(entry.value_type(), valuetype);
         }
     }
 }
